\name{OptionalrSpace-class}
\docType{class}
\alias{OptionalrSpace-class}

\title{Optional rSpace}
\description{Optional object of class \code{"rSpace"}.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
%\section{Methods}{}
%\references{}
\author{Matthias Kohl \email{Matthias.Kohl@stamats.de}}
%\note{}
\seealso{\code{\link[distr]{rSpace-class}}}
\concept{space}
\keyword{classes}
