\name{RandVar-package}
\alias{RandVar-package}
\alias{RandVar}
\docType{package}
\title{
Implementation of random variables
}
\description{
Implementation of random variables by means of S4 classes and methods.
}
\details{
\tabular{ll}{
Package: \tab RandVar \cr
Version: \tab 0.8 \cr
Date: \tab 2010-12-03 \cr
Depends: \tab R (>= 2.7.0), methods, startupmsg, distr(>= 2.0), distrEx(>=
2.0)\cr
LazyLoad: \tab yes \cr
License: \tab LGPL-3 \cr
URL: \tab http://robast.r-forge.r-project.org/\cr
SVNRevision: \tab 439 \cr
}
}
\author{
Peter Ruckdeschel \email{Peter.Ruckdeschel@itwm.fraunhofer.de},\cr%
Matthias Kohl \email{Matthias.Kohl@stamats.de}\cr

Maintainer: Matthias Kohl  \email{matthias.kohl@stamats.de}}
\references{
  M. Kohl (2005). Numerical Contributions to the Asymptotic Theory of Robustness.
  Dissertation. University of Bayreuth.
}
\seealso{
\code{\link[distr:0distr-package]{distr-package}}, \code{\link[distrEx:0distrEx-package]{distrEx-package}}
}
\section{Package versions}{
Note: The first two numbers of package versions do not necessarily reflect
 package-individual development, but rather are chosen for the
 RobAStXXX family as a whole in order to ease updating "depends"
 information.
}

\examples{
library(RandVar)
#vignette("RandVar")
}
\keyword{package}
