\name{RandVar-package}
\alias{RandVar-package}
\alias{RandVar}
\docType{package}
\title{
Implementation of random variables
}
\description{
Implementation of random variables by means of S4 classes and methods.
}
\details{
\tabular{ll}{
Package: \tab RandVar\cr
Version: \tab 0.6.7 \cr
Date: \tab 2009-02-25 \cr
Depends: \tab R (>= 2.7.0), methods, startupmsg, distr(>= 2.0), distrEx(>=
2.0)\cr
LazyLoad: \tab yes\cr
License: \tab LGPL-3\cr
URL: \tab http://robast.r-forge.r-project.org/\cr
}
}
\author{
Peter Ruckdeschel \email{Peter.Ruckdeschel@itwm.fraunhofer.de},\cr%
Matthias Kohl \email{Matthias.Kohl@stamats.de}\cr

Maintainer: Matthias Kohl  \email{matthias.kohl@stamats.de}}
\references{
  M. Kohl (2005). Numerical Contributions to the Asymptotic Theory of Robustness.
  Dissertation. University of Bayreuth.
}
\seealso{
\code{\link[distr]{distr-package}}, \code{\link[distrEx]{distrEx-package}}
}
\examples{
library(RandVar)
#vignette("RandVar")
}
\keyword{package}
