\name{RandVar-package}
\alias{RandVar-package}
\alias{RandVar}
\docType{package}
\title{
Implementation of random variables
}
\description{
Implementation of random variables by means of S4 classes and methods.
}
\details{
\tabular{ll}{
Package: \tab RandVar\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2006-12-14\cr
Depends: \tab R (>= 2.4.0), methods, startupmsg, distr(>= 1.8), distrEx(>= 0.4-4)\cr
License: \tab GPL (version 2 or later)\cr
}
require(RandVar)
}
\author{ Dr. Matthias Kohl

Maintainer: Dr. Matthias Kohl  \email{matthias.kohl@stamats.de}}
\references{
  M. Kohl (2005). Numerical Contributions to the Asymptotic Theory of Robustness.
  Dissertation. University of Bayreuth.
}
%~~ Optionally other keywords from doc/KEYWORDS, one per line
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
\keyword{package}
